feat(iota-types)!: replace CallArg by Input (#10317)
Replaces the local `CallArg` enum in iota-types with a type alias to
`iota_sdk_types::Input`, continuing the effort to unify types between
the IOTA node and the SDK

Fixes https://github.com/iotaledger/iota-rust-sdk/issues/582

---------

+++ crates/iota-grpc-server/src/transaction_execution_service/simulate.rs
@@ -398,7 +398,7 @@ pub async fn simulate_transaction(
                     // Set the command index if available
                     if error_mask.contains(ExecutionError::COMMAND_INDEX_FIELD.name) {
                         if let Some(command_idx) = execution_error.command() {
-                            exec_error.command_index = Some(command_idx as u64);
+                            exec_error.command_index = Some(command_idx);
                         }
                     }
